+++ lib/presentation/widgets/separator.dart
@@ -10,17 +10,16 @@ class Separator extends StatelessWidget {
   Widget build(BuildContext context) {
     //1
     return Container(
-      height: Sizes.dimen_1.h.toDouble(),
-      width: Sizes.dimen_80.w.toDouble(),
+      height: Sizes.dimen_1.h,
+      width: Sizes.dimen_80.w,
       //2
       padding: EdgeInsets.only(
-        top: Sizes.dimen_2.h.toDouble(),
-        bottom: Sizes.dimen_6.h.toDouble(),
+        top: Sizes.dimen_2.h,
+        bottom: Sizes.dimen_6.h,
       ),
       //3
       decoration: BoxDecoration(
-        borderRadius:
-            BorderRadius.all(Radius.circular(Sizes.dimen_1.h.toDouble())),
+        borderRadius: BorderRadius.all(Radius.circular(Sizes.dimen_1.h)),
         gradient: const LinearGradient(
           colors: [
             AppColor.vulcan,
